+++ CHANGELOG.md
@@ -167,3 +167,4 @@
 * Define `Traced` for adaptable `Environment` comonad
 * Define `cata`, `ana` and `hylo` methods of `Fix`
 * Define `Equipment` datatype to use `Comonad` `Product` transformer
+* Define adaptable `retrieve` method of `Equipment`